% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{geeq}
\alias{geeq}
\title{Solve Generalized Estimating Equations and Quadratic Inference Function}
\usage{
geeq(formula, id = NULL, data = parent.frame(), family = gaussian,
  method = "gee", weights = NULL, waves = NULL, maxit = 30,
  tol = 10^-6, corstr = "independence", Mv = 1, cor.mat = matrix(),
  init.alpha = NULL, init.beta = NULL, init.phi = 1, scale.fix = FALSE)
}
\arguments{
\item{formula}{a formula expression similar to that for \code{\link[stats]{glm}}, of the
form \code{response ~ predictors}. A term of the form \code{offset(expression)} is allowed.}

\item{id}{a vector which identifies the clusters. The length of \code{id} should be the same
as the number of observations. Observations with same \code{id} belong to the same cluster.}

\item{data}{an optional data frame containing the variables and \code{id}. \code{weights} and
\code{waves} should also be included if necessary.}

\item{family}{\code{family} argument determine the details of the model. The argument can be
one of three options: a \code{\link[stats]{family}} object, a character string corresponding
to one of the family objects, or a list of user specified functions.

User specified functions must contain following components: \code{linkfun}, \code{linkinv},
\code{variance}, \code{mu.eta} and \code{dev.resids}. See \code{\link[stats]{family}} for more
information.}

\item{method}{a character string specifying the method to estimate. Options could be 'gee' of 'qif'.}

\item{weights}{an optional vector for each observation. Observations with weight 0 are excluded
from the calculation process. All observations will be assigned 1 by default. If an observation
misses any variable, it will be assigned 0.}

\item{waves}{an optional vector identifying time ordering in a cluster.}

\item{maxit}{integer giving the maximal number of iteration.}

\item{tol}{positive convergence tolerance. The iterations converge when the absolute value
of the difference in parameter estimate is below \code{tol}}

\item{corstr}{a character string specifying the correlation structure. Options could be:
'independence', 'fixed', 'ar1', 'exchangeable', 'm-dependent' and 'unstructured'.}

\item{Mv}{for 'm-dependent', the value for \code{m}.}

\item{init.alpha}{an optional vector with initial values of alpha. The length is different for
different correlation structure.}

\item{init.beta}{an optional vector with initial values of beta. If not specificed, then it will
be set to the result of \code{\link[stats]glm}}

\item{init.phi}{an optional initial overdispersion parameter.}

\item{scale.fix}{if set to \code{TRUE}, then the scale parameter is fixed at the value of \code{
init.phi}.}

\item{corr.mat}{the correlations matrix for 'fixed'. Matrix should be symmetric with dimensions
>= the maximum cluster size.}
}
\value{
An object of class 'geeq' representing the fit. Some of the attributes are defined as follows:
\item{\code{beta}}{a named vector of estimated coefficients.}
\item{\code{variance}}{estimated variance of \code{beta}.}
\item{\code{phi}}{estimated scale parameter.}
\item{\code{fitted.values}}{the fitted mean values, obtained by transforming the linear predictors
by the inverse of the link function.}
\item{\code{call}}{the matched call.}
\item{\code{model}}{the model frame.}
\item{\code{offset}}{the priori known component in the linear predictor.}
\item{\code{df.null}}{the residual degrees of freedom for the null model.}
\item{\code{df.residual}}{the residual degrees of freedom.}
\item{\code{null.deviance}}{the deviance for the null model, comparable with deviance. The null model
will include the offset, and an intercept if there is one in the model.}
\item{\code{deviance}}{up to a constant, minus twice the maximized log-likelihood. Where sensible, the constant
is chosen so that a saturated model has deviance zero.}
\item{\code{niter}}{the number of iterations to solve equations.}
\item{\code{converged}}{logical. Whether the algorithm have converged.}
\item{\code{cluster.size}}{a vector of size for each cluster}
\item{\code{ncluster}}{number of clusters}
\item{\code{nobs}}{number of observations from all clusters}

For the result of 'gee' method, the following attributes are provided:
\item{\code{alpha}}{a vector of estimated correlation parameters.}
\item{\code{correlation}}{a matrix of estimated correlation.}
\item{\code{QIC}}{a criteria to select correlation structure.}

As for 'qif' method, the result has following attributes:
\item{\code{Q}}{value of quadratic inference function}
\item{\code{statistics}}{some statistics to evaluate the result}
}
\description{
Calculate coefficients of generalized estimating equations. Both generalised estimating
equations and quadratic inference function method are provided. Produces an object of
class 'geeq' which is a fit of the model.
}
\seealso{
\code{\link[stats]{glm}}, \code{\link[stats]{family}}
}
\author{
Yuze Zhou
}
